Significant speed improvements
Versus my previous code, this pulls out the calls to .forEach(), causing
the code to now take less than half of the time.  I'm a firm believer
that middleware that will be hit millions of times should try to be as
fast as is reasonable, though keeping the code looking good for
maintenance purposes.